🚧 chore: hacks refactoring and documentation
Digging knwoladge back from hacks as well as trying to make it easier to
maintain, understand and modify.

the main goal is to centralise workarounds and knowladge about why they
exists and how they came to be. The secondary goal was to reduce the
amount of state to make is reproductible without rolling back
old-outscale.yaml

Upgrading Ruby from 3.0 to 3.3 in the process, as well as resionalise
shebangs to make it more nix-shell frendly.