♻️refactor: rework patches and breaking change handling
Previous implementation relie on a state `old-outscale.yaml` for doing
type backporting which result in inconsistant implementation in SDKs and
the unability to re-run easeliy the workflow.

This now implementation relis on a stateless aproche and a more modulare
implementaions of hacks. In a ideal world, patches should be declared in
the builder on each SDK and this repositorie whould only be a workflow
dispatcher and a toolbox. However, for backward compabilility of present
workflow, all patches ,`build.sh` and build artifacts (outscale.yaml,
outscale-*.yaml) are still commited to this repositorie for now.

A Workflow have also been added to avoid future bracking chages as well